Refactor using context UNTESTED
