+++ src/components/Header/Header.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useRef } from "react";
+import React, { useEffect, useState, useRef, useContext } from "react";
 import StreamrClient, { Stream, StreamOperation } from "streamr-client";
 import { ethers } from "ethers";
 import { useHistory } from "react-router-dom";
@@ -30,108 +30,59 @@ import {
 import { ChevronDownIcon } from "@chakra-ui/icons";
 
 import logo from "../../assets/streamr.svg";
-import AddModal from './AddModal';
-import { addPermissions } from '../../utils/utils';
+import AddModal from "./AddModal";
+import { addPermissions } from "../../utils/utils";
+import JoinModal from "./JoinModal";
+import { UserContext } from "../../contexts/UserContext";
+import CreateModal from "./CreateModal";
 
 type Props = {
-  setClient: React.Dispatch<React.SetStateAction<StreamrClient>>;
-  address: string;
-  setAddress: React.Dispatch<React.SetStateAction<string>>;
   setProvider: React.Dispatch<
     React.SetStateAction<ethers.providers.Web3Provider>
   >;
-  client: StreamrClient;
-  setConnectedAddress: React.Dispatch<React.SetStateAction<string>>;
-  connectedAddress: string;
 };
 
-const Header = ({
-  setClient,
-  address,
-  setAddress,
-  setProvider,
-  client,
-  setConnectedAddress,
-  connectedAddress,
-}: Props) => {
+const Header = ({ setProvider }: Props) => {
   const { ethereum } = window;
 
   const createDisclosure = useDisclosure();
   const joinDisclosure = useDisclosure();
   const addDisclosure = useDisclosure();
 
-  const [invalidFriendAddress, setInvalidFriendAddress] = useState(false);
-  const [friendAddress, setFriendAddress] = useState("");
   const [code, setCode] = useState<Stream>({} as Stream);
-  const [joinCode, setJoinCode] = useState("");
-  const [rightCode, setRightCode] = useState(false);
-  const [wrongCode, setWrongCode] = useState(false);
-  const [noPermissions, setNoPermissions] = useState(false);
 
   const isFirstRender = useRef(true);
+
+  const {
+    connectedAddress,
+    publicAddress,
+    client,
+    setConnectedAddress,
+    setPublicAddress,
+    setClient,
+  } = useContext(UserContext);
 
   const handleCreate = async () => {
     try {
       const stream = await client.getOrCreateStream({
-        id: `${address.toLowerCase()}/streamr-chat-messages`, // or 0x1234567890123456789012345678901234567890/foo/bar or mydomain.eth/foo/bar
+        id: `${publicAddress.toLowerCase()}/streamr-chat-messages`, // or 0x1234567890123456789012345678901234567890/foo/bar or mydomain.eth/foo/bar
       });
 
       setCode(stream);
-      setConnectedAddress(address.toLowerCase());
+      setConnectedAddress(publicAddress.toLowerCase());
     } catch (err) {
       console.log(err);
       setCode({} as Stream);
     }
   };
 
-  const handleJoin = async () => {
-    const msg = {
-      hello: "world",
-    };
-
-    // Publish the event to the Stream
-    try {
-      await client.publish(joinCode, msg);
-      const words = joinCode.toLowerCase().split("/");
-      setConnectedAddress(words[0]);
-      /* const words = joinCode.split("/");
-      setConnectedAddress(words[0]);
-      const stream = await client.getOrCreateStream({
-        id: `${words[0]}/streamr-chat-messages`, // or 0x1234567890123456789012345678901234567890/foo/bar or mydomain.eth/foo/bar
-      });
-
-      console.log(joinCode);
-
-      if (
-        !(await stream.hasPermission(
-          "stream_publish" as StreamOperation,
-          joinCode
-        ))
-      ) {
-        console.log("no permissions!");
-        throw new Error("You have insufficient permissions!");
-      } else {
-        setRightCode(true);
-        } */
-      setRightCode(true);
-    } catch (err) {
-      console.log(err);
-      /* console.log(err.message);
-      if ((err.message = "You have insufficient permissions!")) {
-        setNoPermissions(true);
-      } else {
-      } */
-      setWrongCode(true);
-    }
-  };
-
   useEffect(() => {
-    console.log(address);
+    console.log(publicAddress);
     if (isFirstRender) {
       isFirstRender.current = false;
       handleCreate();
     }
-  }, [address])
+  }, [publicAddress]);
 
   return (
     <Flex
@@ -148,7 +99,7 @@ const Header = ({
         </Heading>
         <Image src={logo} boxSize="50px" marginLeft="7px" />
         <Spacer />
-        {address ? (
+        {publicAddress ? (
           <>
             <Tooltip label="Click to Copy" placement="bottom">
               <Button
@@ -157,13 +108,20 @@ const Header = ({
                 _hover={{ backgroundColor: "#13013D" }}
                 onClick={() => {
                   handleCreate();
-                  addDisclosure.onOpen()
+                  addDisclosure.onOpen();
                 }}
               >
-                {address}
+                {publicAddress}
               </Button>
             </Tooltip>
-            <AddModal disclosure={addDisclosure} client={client} code={code} address={address} setCode={setCode} setConnectedAddress={setConnectedAddress} />
+            <AddModal
+              disclosure={addDisclosure}
+              client={client}
+              code={code}
+              address={publicAddress}
+              setCode={setCode}
+              setConnectedAddress={setConnectedAddress}
+            />
           </>
         ) : (
           <Button
@@ -172,7 +130,7 @@ const Header = ({
             backgroundColor="#0D009A"
             onClick={async () => {
               if (!ethereum) {
-                setAddress("no wallet detected");
+                setPublicAddress("no wallet detected");
                 return;
               }
               await window.ethereum.enable();
@@ -180,7 +138,7 @@ const Header = ({
               provider.getSigner();
               await provider.send("eth_requestAccounts", []);
               if (!ethereum.selectedAddress) {
-                setAddress("wallet not signed in");
+                setPublicAddress("wallet not signed in");
                 return;
               }
               const client = await new StreamrClient({
@@ -197,10 +155,10 @@ const Header = ({
               } catch (err) {
                 console.log(err);
               }
-              setAddress(ensAddress || ethereum.selectedAddress);
+              setPublicAddress(ensAddress || ethereum.selectedAddress);
               setProvider(provider);
-              setConnectedAddress(address.toLowerCase());
-              addDisclosure.onOpen()
+              setConnectedAddress(publicAddress.toLowerCase());
+              addDisclosure.onOpen();
             }}
           >
             Connect
@@ -223,180 +181,17 @@ const Header = ({
             <MenuItem onClick={joinDisclosure.onOpen}>Join Room</MenuItem>
           </MenuList>
         </Menu>
-        <Modal
-          isOpen={createDisclosure.isOpen}
-          onClose={createDisclosure.onClose}
-          isCentered
-        >
-          <ModalOverlay />
-          <ModalContent>
-            <ModalHeader>Create Room</ModalHeader>
-            <ModalCloseButton />
-            {client ? (
-              <ModalBody>
-                <Input
-                  placeholder="Friend's Public Address"
-                  value={friendAddress}
-                  onChange={(e) => {
-                    setFriendAddress(e.target.value);
-                  }}
-                ></Input>
-              </ModalBody>
-            ) : (
-              <Alert status="error" w="90%" mx="auto" borderRadius="5">
-                You must click connect before creating a room!
-              </Alert>
-            )}
-
-            {invalidFriendAddress && (
-              <Alert status="error" w="90%" mx="auto" borderRadius="5">
-                Invalid Public Address!
-              </Alert>
-            )}
-
-            {code.id && (
-              <Alert
-                status="success"
-                w="90%"
-                mx="auto"
-                borderRadius="5"
-                display="flex"
-                flexDir="column"
-              >
-                Room successfully created! Ask your friend to enter the code:{" "}
-                <Tooltip label="Click to Copy" placement="bottom">
-                  <Box
-                    width="95%"
-                    backgroundColor="#525252"
-                    color="white"
-                    overflow="scroll"
-                    borderRadius="5px"
-                    paddingY="5px"
-                    paddingX="10px"
-                    marginY="5px"
-                    onClick={() => {
-                      navigator.clipboard.writeText(code.id);
-                    }}
-                  >
-                    {code.id}
-                  </Box>
-                </Tooltip>
-              </Alert>
-            )}
-
-            <ModalFooter>
-              {code.id ? (
-                <Button
-                  color="white"
-                  _hover={{ backgroundColor: "#13013D" }}
-                  backgroundColor="#0D009A"
-                  ml={3}
-                  disabled={!client}
-                  onClick={createDisclosure.onClose}
-                >
-                  Ok
-                </Button>
-              ) : (
-                <>
-                  <Button variant="ghost" onClick={createDisclosure.onClose}>
-                    Cancel
-                  </Button>
-                  <Button
-                    color="white"
-                    _hover={{ backgroundColor: "#13013D" }}
-                    backgroundColor="#0D009A"
-                    ml={3}
-                    disabled={!client}
-                    onClick={handleCreate}
-                  >
-                    Confirm
-                  </Button>
-                </>
-              )}
-            </ModalFooter>
-          </ModalContent>
-        </Modal>
-        <Modal
-          isOpen={joinDisclosure.isOpen}
-          onClose={joinDisclosure.onClose}
-          isCentered
-        >
-          <ModalOverlay />
-          <ModalContent>
-            <ModalHeader>Join Room</ModalHeader>
-            <ModalCloseButton />
-            {client ? (
-              <ModalBody>
-                <Input
-                  placeholder="Enter Code"
-                  value={joinCode}
-                  onChange={(e) => {
-                    setRightCode(false);
-                    setWrongCode(false);
-                    setNoPermissions(false);
-                    setJoinCode(e.target.value);
-                  }}
-                ></Input>
-              </ModalBody>
-            ) : (
-              <Alert status="error" w="90%" mx="auto" borderRadius="5">
-                You must click connect before creating a room!
-              </Alert>
-            )}
-
-            {wrongCode && (
-              <Alert status="error" w="90%" mx="auto" borderRadius="5">
-                The code you provided is invalid!
-              </Alert>
-            )}
-
-            {rightCode && (
-              <Alert status="success" w="90%" mx="auto" borderRadius="5">
-                You successfully joined the room!
-              </Alert>
-            )}
-
-            {noPermissions && (
-              <Alert status="error" w="90%" mx="auto" borderRadius="5">
-                You don't have sufficient permissions!
-              </Alert>
-            )}
-
-            <ModalFooter>
-              {rightCode ? (
-                <Button
-                  color="white"
-                  _hover={{ backgroundColor: "#13013D" }}
-                  backgroundColor="#0D009A"
-                  ml={3}
-                  disabled={!client}
-                  onClick={joinDisclosure.onClose}
-                >
-                  Ok
-                </Button>
-              ) : (
-                <>
-                  <Button variant="ghost" onClick={joinDisclosure.onClose}>
-                    Cancel
-                  </Button>
-                  <Button
-                    color="white"
-                    _hover={{ backgroundColor: "#13013D" }}
-                    backgroundColor="#0D009A"
-                    ml={3}
-                    disabled={!client}
-                    onClick={handleJoin}
-                  >
-                    Confirm
-                  </Button>
-                </>
-              )}
-            </ModalFooter>
-          </ModalContent>
-        </Modal>
+        <CreateModal
+          disclosure={createDisclosure}
+          handleCreate={handleCreate}
+          code={code}
+          setCode={setCode}
+        />
+        <JoinModal disclosure={joinDisclosure} client={client} />
       </Flex>
-      <Text marginTop="10px">{`Room: ${connectedAddress || "Not in Room"
-        }`}</Text>
+      <Text marginTop="10px">{`Room: ${
+        connectedAddress || "Not in Room"
+      }`}</Text>
     </Flex>
   );
 };
